Support pip uninstall. Important note:
The scripts install to the base directory of the active python
installation. So if `pip` found on your path is actually
`C:\Python27\Scripts\pip.exe`, the scripts will be installed
in `C:\Python27`. Previously they were installed in the Scripts
directory; I've moved them to the python root directory for 2 reasons:

1. For some people the root python directory is the only python
directory they want on their PATH. (It contains python.exe and pythonw.exe)

2. If you're using a python switcher, such as `py` or my own
`[pywin](https://github.com/davidmarble/pywin)`, when you switch python
versions it's bad news to have more than one `Scripts\` directory from
separate python versions on the PATH. So taking `pywin` as an example,
when you switch python versions I leave the default python root directory
on the PATH, but remove the Scripts directory if it's on the PATH, then
I put the new python version's root and Scripts directories at the front
of the PATH.

Illustrating further... If your PATH starts out as

    C:\Windows;C:\Python27;C:\Python27\Scripts

and you run my pywin launcher:

    C:\> pywin setdefault 3.3

the PATH will now be:

    C:\Python33;C:\Python33\Scripts;C:\Windows;C:\Python27

Note that `C:\Python27\Scripts` is no longer on the PATH. So it's best if
virtualenvwrapper-win scripts are in `C:\Python27`, which will always be
on the PATH.